Revert adding Relationship to project clients table
